Added ability to remove rows and boxes.
Added a minus box at the end of each row and a delete row button
at the top. Also made the add box function simpler and changed the
inside of the click event to a switch statement to make it clearer.
Altered the name of the style class to reflect the change and
included the new box in the HTML.